+++ docs/app/layouts/docs.vue
@@ -35,13 +35,8 @@ const { mobileLinks } = useHeader()
       <B24SidebarHeader>
         <LogoWithVersion />
       </B24SidebarHeader>
-      <B24SidebarBody
-        :b24ui="{
-          root: '[&>[data-slot=section]+[data-slot=section]]:mt-0'
-        }"
-      >
+      <B24SidebarBody>
         <B24NavigationMenu
-          class="lg:hidden"
           :items="mobileLinks"
           orientation="vertical"
         />