+++ docs/app/pages/icons/index.vue
@@ -16,14 +16,32 @@ useSeoMeta({
   description,
   ogDescription: description
 })
+
+const { mobileLinks } = useHeader()
 </script>
 
 <template>
   <B24SidebarLayout
     :use-light-content="false"
   >
+    <template #sidebar>
+      <B24SidebarHeader>
+        <LogoWithVersion />
+      </B24SidebarHeader>
+      <B24SidebarBody>
+        <B24NavigationMenu
+          :items="mobileLinks"
+          orientation="vertical"
+        />
+      </B24SidebarBody>
+      <B24SidebarFooter>
+        <B24SidebarSection>
+          <ExtLinks />
+        </B24SidebarSection>
+      </B24SidebarFooter>
+    </template>
     <template #navbar>
-      <Header show-logo-all-time />
+      <Header />
     </template>
 
     <IconsList :groups="groups" />