+++ docs/server/api/component-example.get.ts
@@ -1,0 +1,21 @@
+import { defineEventHandler, createError, appendHeader } from 'h3'
+import { pascalCase } from 'scule'
+// @ts-expect-error - no types available
+import components from '#component-example/nitro'
+
+export default defineEventHandler((event) => {
+  appendHeader(event, 'Access-Control-Allow-Origin', '*')
+
+  const componentName = (event.context.params?.['component?'] || '').replace(/\.json$/, '')
+
+  if (componentName) {
+    const component = components[pascalCase(componentName)]
+    if (!component) {
+      throw createError({
+        statusMessage: 'Example not found!',
+        statusCode: 404
+      })
+    }
+    return component
+  }
+})